Refactor address page with library user interface
